+++ docs/src/releasenotes.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## Unreleased
+* Adds `pyhasitem` and 3-arg `pygetitem`.
+* Extends `Base.get`, `Base.get!`, `Base.haskey` and 2-arg `Base.hash` for `Py`.
+* `PyArray` can now have any element type when the underlying array is of Python objects.
+* Bug fixes.
+
 ## v0.8.0 (2022-03-17)
 * **Breaking:** Removes `pymethod` and `pyclass`. In the future, `pyclass` may become sugar
   for `types.new_class` (namely you can specify a metaclass).